B-Splines changed to operate on doubles, First part of FilamentSegmentation algorithm finished (minimizing energy), added JUnit tests

+++ plugin/src/mosaic/nurbs/BSplineSurfaceFactory.java
@@ -19,7 +19,7 @@ public class BSplineSurfaceFactory {
 	 * @param aFunc
 	 * @return
 	 */
-	public static BSplineSurface generateFromFunction(float aUmin, float aUmax, float aVmin, float aVmax, float aScale, int aDegree, Function aFunc) {
+	public static BSplineSurface generateFromFunction(double aUmin, double aUmax, double aVmin, double aVmax, double aScale, int aDegree, Function aFunc) {
 		// Calculate number of steps in each directions and create container for function values
 		int noOfStepsU = (int)((aUmax - aUmin) / aScale) + 1;		
 		int noOfStepsV = (int)((aVmax - aVmin) / aScale) + 1;
@@ -31,7 +31,7 @@ public class BSplineSurfaceFactory {
 	 * Generates surface in given range in u/v directions basing on provided function and numbers of steps.
 	 * @return B-Spline surface
 	 */
-	public static BSplineSurface generateFromFunction(float aUmin, float aUmax, float aVmin, float aVmax, int aNoOfSteps, int aDegree, Function aFunc) {
+	public static BSplineSurface generateFromFunction(double aUmin, double aUmax, double aVmin, double aVmax, int aNoOfSteps, int aDegree, Function aFunc) {
 		return generateFromFunction(aUmin, aUmax, aVmin, aVmax, aNoOfSteps, aNoOfSteps, aDegree, aDegree, aFunc);
 	}
 		
@@ -40,19 +40,19 @@ public class BSplineSurfaceFactory {
 	 * be configured independently in each direction.
 	 * @return B-Spline surface
 	 */
-	public static BSplineSurface generateFromFunction(float aUmin, float aUmax, float aVmin, float aVmax, int aNoOfStepsInUdir, int aNoOfStepsInVdir, int aDegreeInUdir, int aDegreeInVdir, Function aFunc) {
-		float[][] func = new float[aNoOfStepsInUdir][aNoOfStepsInVdir];
+	public static BSplineSurface generateFromFunction(double aUmin, double aUmax, double aVmin, double aVmax, int aNoOfStepsInUdir, int aNoOfStepsInVdir, int aDegreeInUdir, int aDegreeInVdir, Function aFunc) {
+		double[][] func = new double[aNoOfStepsInUdir][aNoOfStepsInVdir];
 
 		// Calculate step value in each direction
-		float iStepU = (aUmax - aUmin) / (aNoOfStepsInUdir - 1);
-		float iStepV = (aVmax - aVmin) / (aNoOfStepsInVdir - 1);
+		double iStepU = (aUmax - aUmin) / (aNoOfStepsInUdir - 1);
+		double iStepV = (aVmax - aVmin) / (aNoOfStepsInVdir - 1);
 
 		// Calculate values of function in for all requested points
 		for (int u = 0; u < aNoOfStepsInUdir; ++u) {
 			for (int v = 0; v < aNoOfStepsInVdir; ++v) {
-				float uVal = u * iStepU + aUmin;
-				float vVal = v * iStepV + aVmin;
-				float val = aFunc.getValue(uVal, vVal);
+				double uVal = u * iStepU + aUmin;
+				double vVal = v * iStepV + aVmin;
+				double val = aFunc.getValue(uVal, vVal);
 				
 				func[u][v] = val;
 			}
